+++ tests/aleph_client/src/lib.rs
@@ -1,3 +1,11 @@
+mod psp22_contract;
+mod test_contract;
+
+#[cfg(test)]
+mod psp22_tests;
+#[cfg(test)]
+mod test_contract_tests;
+
 use std::sync::Mutex;
 
 use aleph_client::{